rename trustification to bom-dependencies and reduce executions
