+++ src/test/resources/bom-dependencies/test_nested_dependency3/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.example</groupId>
+        <artifactId>bom-dependencies-parent</artifactId>
+        <version>1.0.0</version>
+    </parent>
+
+    <artifactId>test_nested_dependency3</artifactId>
+
+    <name>Test Nested Dependency 3</name>
+</project>
