--- src/test/resources/issue-trustification1/pom.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>com.example</groupId>
-    <artifactId>trustification-parent</artifactId>
-    <packaging>pom</packaging>
-    <version>1.0.0</version>
-
-    <name>Trustification Parent</name>
-
-    <modules>
-        <module>trustification</module>
-
-        <module>versioned_dependency1</module>
-        <module>versioned_dependency2</module>
-        <module>provided_dependency</module>
-
-        <module>dependency1</module>
-        <module>dependency2</module>
-
-        <module>test_dependency</module>
-        <module>test_compile_dependency</module>
-        <module>test_nested_dependency1</module>
-        <module>test_nested_dependency2</module>
-        <module>test_nested_dependency3</module>
-
-        <module>shared_dependency1</module>
-        <module>shared_dependency2</module>
-        <module>shared_runtime_dependency</module>
-
-        <module>runtime_dependency</module>
-        <module>shared_runtime_dependency1</module>
-        <module>shared_runtime_dependency2</module>
-
-        <module>type_dependency</module>
-        <module>shared_type_dependency1</module>
-        <module>shared_type_dependency2</module>
-        <module>shared_type_dependency3</module>
-        <module>shared_type_dependency4</module>
-    </modules>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.cyclonedx</groupId>
-                <artifactId>cyclonedx-maven-plugin</artifactId>
-                <version>${test.input.version}</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>makeBom</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <projectType>library</projectType>
-                    <schemaVersion>1.3</schemaVersion>
-                    <includeBomSerialNumber>true</includeBomSerialNumber>
-                    <includeCompileScope>true</includeCompileScope>
-                    <includeProvidedScope>true</includeProvidedScope>
-                    <includeRuntimeScope>false</includeRuntimeScope>
-                    <includeSystemScope>false</includeSystemScope>
-                    <includeTestScope>false</includeTestScope>
-                    <includeLicenseText>true</includeLicenseText>
-                    <outputFormat>all</outputFormat>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-</project>
