--- src/test/resources/issue-trustification1/test_nested_dependency2/pom.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.example</groupId>
-        <artifactId>trustification-parent</artifactId>
-        <version>1.0.0</version>
-        <relativePath>..</relativePath>
-    </parent>
-
-    <groupId>com.example</groupId>
-    <artifactId>test_nested_dependency2</artifactId>
-    <packaging>jar</packaging>
-    <version>1.0.0</version>
-
-    <name>Test Nested Dependency 2</name>
-
-    <dependencies>
-      <dependency>
-        <groupId>com.example</groupId>
-        <artifactId>test_nested_dependency3</artifactId>
-        <version>1.0.0</version>
-        <scope>compile</scope>
-      </dependency>
-    </dependencies>
-</project>
